+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### master
+
+- Support for `archive()` and `unarchive()` on `RawRepresentable` types
+
 ### 2.1.3 (2016-03-02)
 
 - Fix Carthage build